[C8] Split serialize_dressup.go per feature (#2237)
model/feature/serialize_dressup.go was 545 LOC, over the 500-line ceiling.
Split the dress-up feature serialization per feature; each file is now under
the ceiling:

- serialize_dressup.go (368): the shared edge/face reference + anchor encoding
  (anchor/geom-edge/geom-face/key encode-decode, require*Dress, evalFloat) plus
  the chamfer and thread data and the misc dress-up DTOs.
- serialize_dressup_fillet.go (143): the fillet data — edge sets (constant/
  variable radius + intermediate points), corner/cross-section modes, the
  face/rule/full-round variants and their serialize/restore.
- serialize_dressup_draft.go (55): the shell/draft data — draft pull-direction
  and neutral-plane encoding, shell per-face thickness.

Behaviour-preserving: split programmatically by receiver-qualified
declaration, verified every non-comment code line is byte-identical (only
import redistribution differs). serialize_*.go is already funlen/gocyclo-
excluded by path in .golangci.yml, so no config change is needed. Feature
package builds, archguard passes, lint 0. Net delta of recognizers / tolerance
constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2237.

(cherry picked from commit 97f190592721f6285733dbb4237224ef30650404)
